% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{availableDataSets}
\alias{availableDataSets}
\title{Load available microbiome data sets}
\usage{
availableDataSets()
}
\value{
A \code{data.frame} containing the following columns:
\itemize{
\item{Dataset:} {the name of the function to load a dataset}
}
}
\description{
To list the available datasets in \code{microbiomeDataSets},
run \code{availableDataSets}.
}
\details{
For information visit the individual man pages.
}
\examples{
availableDataSets()
}
