% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMPCData.R
\name{HMPCData}
\alias{HMPCData}
\alias{HMPCV13Data}
\alias{HMPCV35Data}
\title{Obtain the Human Microbiome Project Consortium dataset}
\usage{
HMPCV13Data()

HMPCV35Data()
}
\value{
a \linkS4class{TreeSummarizedExperiment}
}
\description{
Obtain the selected microbiome datasets from Human Microbiome Project
Consortium publication (2012).
}
\details{
The HMPC dataset comes in two different flavours, both based on 16s rRNA
sequencing using pyrosequencing. HMPCV13Data contains information generated
from sequencing the variable regions 1-3, whereas HMPCV35Data contains
the information extracted from sequencing region 3-5.

Column metadata includes among others information on body sites and
extraction methods used.

Row metadata contains taxonomic information on the Phylum, Family and Genus
level.

All data are downloaded from ExperimentHub and cached for local re-use.
}
\examples{
tse <- HMPCV13Data()
}
\references{
Human Microbiome Project Consortium. Structure, function and diversity of the
healthy human microbiome. Nature. 2012 Jun 13;486(7402):207-14. doi:
10.1038/nature11234. PMID: 22699609; PMCID: PMC3564958.
}
\author{
Felix G.M. Ernst
}
