% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiomeDataSets.R
\name{microbiomeDataSets-package}
\alias{microbiomeDataSets-package}
\title{Microbiome data sets}
\description{
\code{microbiomeDataSets} is a collation of data from microbiome and
associated studies, which are publicaly available.
}
\details{
The data is made available through the ExperimentHub resources of the
Bioconductor project. It is loaded as \code{TreeSummarizedExperiment}
object or a \code{MultiAssayExperiment} objects.
}
\section{Deprecated functions}{

The following functions are deprecated and will be removed in a future
release:
\itemize{
\item \code{LahtiWAdata}: Please use the \code{data()} function from the
\code{mia} package to load this dataset.
\item \code{HintikkaXOData}: Please use the \code{data()} function from
the \code{miaViz} package to load this dataset.
}
}

