% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OKeefeDSData.R
\name{OKeefeDSData}
\alias{OKeefeDSData}
\title{Obtain the O'Keefe diet swap microbiome data}
\usage{
OKeefeDSData()
}
\value{
a \linkS4class{TreeSummarizedExperiment}
}
\description{
Obtain the microbiome data from O'Keefe et al. (2015).
}
\details{
The OKeefeDS dataset contains microbiome data from a study with African and
African American groups undergoing a two-week diet swap.

Column metadata includes the subject identifier, sex, nationality, group
information, sample identifier, time point information, time point
information within group and BMI group.

Row metadata contains taxonomic information on the Phylum, Family and Genus
level.

All data are downloaded from ExperimentHub and cached for local re-use.
}
\examples{
tse <- OKeefeDSData()
}
\references{
O'Keefe S et al. (2015): Fat, fibre and cancer risk in African Americans and
rural Africans. \emph{Nature Communications} 6:6342 (2015)
\url{https://dx.doi.org/10.1038/ncomms7342}
}
\author{
Felix G.M. Ernst
}
