% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SailaniIPOPData.R
\name{SailaniIPOPData}
\alias{SailaniIPOPData}
\title{Obtain the Sailani deep longitudinal multiomics data}
\usage{
SailaniIPOPData()
}
\value{
for  \code{SailaniIPOPData} a \linkS4class{MultiAssayExperiment} object
consisting of two \linkS4class{TreeSummarizedExperiment} objects and
five \linkS4class{SummarizedExperiment} objects.
}
\description{
Obtain the deep longitudinal multiomics data from Sailani et al. (2020).
}
\details{
The SailaniIPOP dataset contains longitudinal multiomics data from profiling
of 105 generally healthy individuals. This dataset consists of 16S rRNA gene
based profiling of gut and nasal microbiota. Host transcriptome, proteome and
metabolome from plasma, cytokine profile, and associated clinical data. The
samples were collected over a period of 4 years. The data set can be used
to investigate longitudinal dynamics and interaction between microbiota and
host associated factors.

This microbiota data as currently available as relative abundances. Hierarchical
taxonomic information is unavailable for microbiota data.

Column metadata includes the time points, sex, subject identifier, sample
identifier and treatment group.

All data are downloaded from ExperimentHub and cached for local re-use.
}
\examples{
mae <- SailaniIPOPData()
#tse <- SailaniIPOPData()
}
\references{
Sailani, M.R., Metwally, A.A., Zhou, W. et al. (2020) Deep longitudinal multiomics
profiling reveals two biological seasonal patterns in California.
\emph{Nat Commun.} 11, 4933 \url{https://doi.org/10.1038/s41467-020-18758-1}

Zhou, W., Sailani, M.R., Contrepois, K. et al. (2019) Longitudinal multi-omics of
host-microbe dynamics in prediabetes.
\emph{Nature} 569, 663-671 \url{https://doi.org/10.1038/s41586-019-1236-x}
}
\author{
Sudarshan Shetty and Felix G.M. Ernst
}
