% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HintikkaXOData.R
\name{HintikkaXOData}
\alias{HintikkaXOData}
\title{Retrieve HintikkaXO data}
\usage{
HintikkaXOData()
}
\value{
for  \code{HintikkaXOData} a \linkS4class{MultiAssayExperiment}
is provided.
}
\description{
Obtain microbiome, metabolite and biomarker data from Hintikka et al. (2021).
}
\details{
The HintikkaXO dataset contains high-throughput profiling data from 40 rat
samples, including 39 biomarkers, 38 metabolites (NMR), and 12706 OTUs from
318 species, measured from Cecum. This is diet comparison study with
High/Low fat diet and xylo-oligosaccaride supplementation.

Column metadata includes the SampleID, RatID, Measurement site, Diet group,
Diet Fat (Low/High),  Diet Supplement (XOS 0/1)

Row metadata of the microbiome data contains taxonomic information on the
Phylum, Class, Order, Family, Genus, Species, and OTU levels.

Biomarker and NMR metabolite data are provided through the
\linkS4class{MultiAssayExperiment} mechanism.

Biomarker data contains 39 biomarkers.

NMR data contains 38 metabolites.

All data are downloaded from ExperimentHub and cached for local re-use.
}
\examples{
# tse <- HintikkaXOData()

}
\references{
Hintikka L et al. (2021): Xylo-oligosaccharides in prevention of hepatic
steatosis and adipose tissue inflammation: associating taxonomic and
metabolomic patterns in fecal microbiomes with biclustering.
\emph{International Journal of Environmental Research and Public Health}
18(8):4049 \url{https://doi.org/10.3390/ijerph18084049}
}
\author{
Leo Lahti
}
